requirements.txt
aiosqlite==0.19.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.3.0
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.104.1
greenlet==3.2.2
h11==0.16.0
idna==3.10
passlib==1.7.4
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.27
starlette==0.27.0
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.24.0
